Updated middleware
